front: Split Discover's view along what its parts answer to.
`DiscoverExperienceView.tsx` stood at 1213 lines, past the 800 the
development guide calls "split now" — the finding a reviewer raised on
#561 and #562 was opened for. Nothing here changes what a reader sees:
every line moved is the line that was there.

Split by what each part answers to, not by size:

- `useDiscoverMap.ts` — the MapLibre instance and every listener on it,
  built once on mount. It answers to the map's lifetime and touches React
  only through the refs handed to it, which is why the whole block moved
  without a seam.
- `useDiscoverHover.ts` — hover, in both directions. Not a slice of state
  but the join between the two surfaces: a row hovered rings every point
  its object is drawn at, a marker hovered scrolls the row into view, and
  both read one flag saying which side the hover came from. Apart, that
  flag would have to become two.
- `DiscoverExperienceList.tsx` — the header, the name filter and the rows,
  presentational throughout.

What stays is what only the view can hold: the marker data it feeds the
source, and every camera move — the one thing a reader can undo by moving
the map themselves, so the guards that tell those apart belong beside it.
686 lines.

The docblocks carry the rules each file must keep, because most of them
were a defect once: a click testing `properties.folded` rather than a
pin's shape, a ring covering every pin of a serial site, a cluster answer
checked against a generation because `getClusterLeaves` resolves after
the pointer has moved on. Four doc pointers now name the file that holds
the thing rather than the file that used to, `AGENTS.md` among them — it
carried the same claim about Discover's inline map style a third time.

`refactoring-plan.md`'s open item 7 named the old file for the
scroll-to-centre it lists as duplicated; it now names
`discover/useDiscoverHover.ts`, so whoever picks that item up greps a file
that holds the code. Item 5 keeps its stale name deliberately — that one
was already pointing at the wrong file on main. `DISC-03` in the tracked
testing matrix named the old file for hover sync, and `DISC-02` for the
cluster click that moved with the listeners; both now name the file that
holds it.

The two icon-only buttons this moved — the list's back arrow and the
filter's clear — gained an `aria-label` on the way, since a screen reader
announced each as "button". They are 2 of 130 unlabelled `IconButton`s in
the frontend, and the sweep with its convention is #564 rather than a
partial pass hidden in a refactor.

Closes #562
